Move out of OneSignal.h
* Into OneSignalNotificationSettings.h:
  - OSNotificationPermission
  - OSPermissionState
  - OSUserResponseBlock
- (This also means OneSignalNotificationSettings.h doesn't need to import OneSignal.h)
- OSNotificationsManager.h will import "OneSignalNotificationSettings.h"
