+++ iOS_SDK/OneSignalSDK/Source/OneSignal.h
@@ -114,34 +114,6 @@
 - (void)onDidDismissInAppMessage:(OSInAppMessage *)message;
 @end
 
-typedef NS_ENUM(NSInteger, OSNotificationPermission) {
-    // The user has not yet made a choice regarding whether your app can show notifications.
-    OSNotificationPermissionNotDetermined = 0,
-    
-    // The application is not authorized to post user notifications.
-    OSNotificationPermissionDenied,
-    
-    // The application is authorized to post user notifications.
-    OSNotificationPermissionAuthorized,
-    
-    // the application is only authorized to post Provisional notifications (direct to history)
-    OSNotificationPermissionProvisional,
-    
-    // the application is authorized to send notifications for 8 hours. Only used by App Clips.
-    OSNotificationPermissionEphemeral
-};
-
-// Permission Classes
-@interface OSPermissionState : NSObject
-
-@property (readonly, nonatomic) BOOL reachable;
-@property (readonly, nonatomic) BOOL hasPrompted;
-@property (readonly, nonatomic) BOOL providesAppNotificationSettings;
-@property (readonly, nonatomic) OSNotificationPermission status;
-- (NSDictionary* _Nonnull)toDictionary;
-
-@end
-
 @interface OSPermissionStateChanges : NSObject
 
 @property (readonly, nonnull) OSPermissionState* to;
@@ -314,7 +286,6 @@ NS_SWIFT_NAME(login(externalId:token:));
 + (void)onesignalLog:(ONE_S_LOG_LEVEL)logLevel message:(NSString* _Nonnull)message;
 
 #pragma mark Prompt For Push
-typedef void(^OSUserResponseBlock)(BOOL accepted); // moved 🔔
 
 + (void)promptForPushNotificationsWithUserResponse:(OSUserResponseBlock)block; // moved 🔔
 + (void)promptForPushNotificationsWithUserResponse:(OSUserResponseBlock)block fallbackToSettings:(BOOL)fallback; // moved 🔔
